// ui/src/api/models/UnitsEnum.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export enum UnitsEnum {
    TSP = 'tsp',
    TBSP = 'tbsp',
    FL_OZ = 'fl oz',
    C = 'c',
    PT = 'pt',
    QT = 'qt',
    GAL = 'gal',
    ML = 'ml',
    L = 'l',
    G = 'g',
    KG = 'kg',
    OZ = 'oz',
    LB = 'lb',
    PINCH = 'pinch',
    DASH = 'dash',
    TOUCH = 'touch',
    HANDFUL = 'handful',
    STICK = 'stick',
    CAN = 'can',
    PKG = 'pkg',
    JAR = 'jar',
    BOTTLE = 'bottle',
    BUNCH = 'bunch',
    SLICE = 'slice',
    PIECE = 'piece',
    WHOLE = 'whole',
    HALF = 'half',
    QUARTER = 'quarter',
    DROP = 'drop',
    CUBE = 'cube',
    TO_TASTE = 'to taste',
}